evaluation/requirements.txt
python-dotenv
google-generativeai
deepl
nltk